#914 #919 An f64 world, all eight spatial tiers, and every query shape that reads them
WriteSpatial threw for every spatial field type except AABB2F, so a 3D archetype fell back to the MVCC path and paid one WAL frame per entity per tick; the f64 tiers were declared in the schema, rejected at ConfigureSpatialGrid, and threw on query with a message pointing at a closed issue. Both halves are gone. The write barrier specialises all eight SpatialFieldType variants through one shared ApplySpatialWrite whose is3D is a call-site constant, and the AABB, radius, ray, frustum and kNN query shapes take f64 world coordinates end to end.

The world frame is f64 — SpatialGridConfig's bounds and cell size, every cell origin derived from them, and the persisted config record, which outgrew BootstrapDictionary.Int6 exactly as the design predicted and now reads both the old 8-int shape and the new 16-int one. Storage did not widen: C15 keeps every stored cluster bound f32 and CELL-RELATIVE, which is precisely what lets a 24-bit mantissa span an f64 world, and widening it would have halved R-Tree fanout and the broadphase's batch width for nothing.

Both traps the original deferral left behind are closed. MaybeFlagMigration receives the real Z centre instead of the hard-coded zero, so write-time and fence-time no longer place the same entity in different Z planes while every counter balances; and the outlier guard's Z axis, documented unreachable since #872, executes for the first time along with the shrink mask's 0x10/0x20 bits. MaybeFlagMigration also moved to a cell-relative f32 test, so the f64 origin costs the hot write path three subtractions rather than six f64 comparisons.

Nine sites spelled "is this archetype 3D?" as FieldType == AABB3F || FieldType == BSphere3F. That was exactly right while the f64 tiers were rejected and became a silent misclassification the moment they were not: an AABB3D archetype reads as 2D, the cell walk pins to FlatPlaneZ, and every entity anywhere else on Z is never visited. SpatialFieldTypeExtensions.Is3D replaced all nine. Rule SQ-06 records that, the two frames the query runs in and the width each one is entitled to.

The narrowing at the boundaries went with them. EcsQuery, SpatialTriggerSystem and SpatialInterestSystem each took doubles from the caller and cast them to float on the way into the index — a round trip through a width neither end asked for. A trigger volume quantised at 10^9 fires for entities outside it and stays silent for entities inside, with no error either way; a quantised ray DIRECTION accumulates error with distance rather than staying bounded; a quantised frustum bounding box collapses the cell range and skips whole cells of the view.

An f32-tier archetype on a world f32 cannot address now fails at InitializeArchetypes rather than filing entities into the wrong cells. The line is that one f32 step at the world's extreme must be finer than a cell — deliberately not #919 AC-9's literal "every cell origin exactly representable in f32", which was implemented first and rejects a world spanning 0.1 to 1000.1 with 100-unit cells that resolves four orders finer than it needs to. Failing startup for a world that works is the worse bug, and the fractional-bounds case is pinned as a regression test.

ClusterRef.SpatialBounds returns a new ClusterWorldAabb of six doubles instead of world values written into the cell-relative ClusterSpatialAabb, where one struct meant two incompatible things depending on which side of the property you held it. The three f32 converters that lost their last production caller when ray and kNN went f64 are deleted rather than left as ways to quantise a world coordinate on the public surface of the type whose job is the opposite. ClusterSpatialQueryResult carries f64 bounds for every tier — exact for an f32 archetype, and the only way an f64 one can see the coordinate its component actually holds.

The f32 path is measured, interleaved against the pre-change build across four pairs on an M4: a 3x3-cell query 423.5 -> 424.3 ns, with the per-cell and narrowphase terms inside the run-to-run spread. Getting there needed one fix that only the measurement could have found. The two new f64 dispatch branches took ClusterSpatialQuery.AABB<TBox> from 716 to 814 bytes of IL, and the JIT sizes an inlining candidate before folding the typeof(TBox) tests — so the specialised body stayed small while the method stopped being inlined, and since it returns a 1.6 KB ref struct by value that cost +44 ns on every query, fixed rather than per-cell. The attribute that takes it back carries the warning for whoever adds the next box variant.

+++ demo/SpaceBattle/Program.cs
@@ -204,7 +204,7 @@ public static class Program
                 continue;
             }
             var a = cluster.SpatialBounds;
-            if (!(a.MinX <= a.MaxX) || float.IsInfinity(a.MinX))
+            if (!(a.MinX <= a.MaxX) || double.IsInfinity(a.MinX))
             {
                 continue;
             }
@@ -213,7 +213,7 @@ public static class Program
             {
                 ArchetypeId = host.ShipArchetypeId,
                 ChunkId = cluster.ChunkId,
-                MinX = a.MinX, MinY = a.MinY, MaxX = a.MaxX, MaxY = a.MaxY,
+                MinX = (float)a.MinX, MinY = (float)a.MinY, MaxX = (float)a.MaxX, MaxY = (float)a.MaxY,
                 LiveCount = cluster.LiveCount,
                 HomeCellKey = home,
                 CentreCellKey = host.Grid.WorldToCellKey(0.5f * (a.MinX + a.MaxX), 0.5f * (a.MinY + a.MaxY), 0f),
